# backend/requests/openapi.yaml
openapi: 3.1.0
info:
  title: Instant Wellness Kits API
  description: |
    REST API for order management with automatic tax calculation and geocoding.

    **Key Features:**
    - 📍 Geocoding with Mapbox (New York State only)
    - 💰 Automatic tax calculation (state, county, city, MCTD)
    - 📊 Async CSV batch import with real-time progress
    - 📡 SSE (Server-Sent Events) for live updates
    - 📈 Detailed error reporting and job history

    **Tech Stack:**
    - Spring Boot 4.0.3
    - Spring Batch 6.0
    - PostgreSQL + Redis
    - Mapbox Geocoding API
  contact:
    name: Acheron Team
    email: support@acheron.com
  version: "1.0"
servers:
- url: http://localhost:8080
  description: Local development server
- url: https://api.acheron.com
  description: Production server
...
paths:
...
  /api/v1/orders/import:
    post:
      tags:
      - Batch Import
      summary: Start async CSV import
      description: Queues a CSV file for asynchronous batch processing. Returns immediately
        with job ID for tracking progress via SSE.
      operationId: startImport
      requestBody:
        content:
          multipart/form-data:
            schema:
              type: object
              properties:
                file:
                  type: string
                  format: binary
                  description: "CSV file with columns: latitude, longitude, subtotal,\
                    \ timestamp"
              required:
              - file
      responses:
        "200":
          description: Import job queued successfully
          content:
            '*/*':
              schema:
                type: string
                example:
                  message: Import job queued successfully
                  jobId: 1
                  filename: orders.csv
                  status: QUEUED
        "400":
          description: Invalid file or upload failed
...
  /api/v1/orders/import/{jobId}:
    get:
      tags:
      - Batch Import
      summary: Get job details
      description: "Retrieves detailed information about a specific batch import job\
        \ including status, metrics, and timing"
      operationId: getJobDetails
      parameters:
      - name: jobId
        in: path
        description: Job ID
        required: true
        schema:
          type: integer
          format: int64
      responses:
        "200":
          description: Job details retrieved
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/BatchJobResponse"
        "404":
          description: Job not found
  /api/v1/orders/import/{jobId}/report:
    get:
      tags:
      - Batch Import
      summary: Get complete job report
      description: Retrieves comprehensive job report including job details and all
        error records. Suitable for export as JSON.
      operationId: getJobReport
      parameters:
      - name: jobId
        in: path
        description: Job ID
        required: true
        schema:
          type: integer
          format: int64
      responses:
        "200":
          description: Report generated successfully
          content:
            '*/*':
              schema:
                type: string
                example: "{\"job\": {...}, \"errors\": [...]}"
        "404":
          description: Job not found
  /api/v1/orders/import/{jobId}/errors:
    get:
      tags:
      - Batch Import
      summary: Get job errors
      description: "Retrieves all error details for failed records in a specific job.\
        \ Includes coordinates, error type, and messages."
      operationId: getJobErrors
      parameters:
      - name: jobId
        in: path
        description: Job ID
        required: true
        schema:
          type: integer
          format: int64
      responses:
        "200":
          description: Errors retrieved successfully
          content:
            '*/*':
              schema:
                type: array
                items:
                  type: object
                  additionalProperties: {}
        "404":
          description: Job not found
  /api/v1/orders/import/status/{jobId}:
    get:
      tags:
      - Batch Import
      summary: Stream job status via SSE
      description: Establishes Server-Sent Events connection for real-time job progress
        updates. Auto-closes when job completes or fails. Updates every 1 second.
      operationId: streamJobStatus
      parameters:
      - name: jobId
        in: path
        description: Job ID returned from start import
        required: true
        schema:
          type: integer
          format: int64
      responses:
        "200":
          description: SSE stream established
          content:
            text/event-stream: {}
        "404":
          description: Job not found
  /api/v1/orders/import/history:
    get:
      tags:
      - Batch Import
      summary: Get job history
      description: Retrieves paginated list of all batch import jobs ordered by creation
        date (newest first)
      operationId: getJobHistory
      parameters:
      - name: page
        in: query
        description: Page number (0-indexed)
        required: false
        schema:
          type: integer
          format: int32
          default: 0
      - name: size
        in: query
        description: Page size (max 100)
        required: false
        schema:
          type: integer
          format: int32
          default: 20
      responses:
        "200":
          description: Job history retrieved successfully
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/PagedModel"
...
components:
  schemas:
...
    BatchJobResponse:
      type: object
      properties:
        jobId:
          type: integer
          format: int64
        jobName:
          type: string
        filename:
          type: string
        status:
          type: string
        startTime:
          type: string
          format: date-time
        endTime:
          type: string
          format: date-time
        totalRecords:
          type: integer
          format: int32
        processedRecords:
          type: integer
          format: int32
        successfulRecords:
          type: integer
          format: int32
        failedRecords:
          type: integer
          format: int32
        skippedRecords:
          type: integer
          format: int32
        errorMessage:
          type: string
        durationMs:
          type: integer
          format: int64
    PageMetadata:
      type: object
      properties:
        size:
          type: integer
          format: int64
        number:
          type: integer
          format: int64
        totalElements:
          type: integer
          format: int64
        totalPages:
          type: integer
          format: int64
    PagedModel:
      type: object
      properties:
        content:
          type: array
          items: {}
        page:
          $ref: "#/components/schemas/PageMetadata"